refactor: implement discoverer for aws provider
