+++ provider/v2/aws/provider copy.go
@@ -84,49 +84,6 @@ func (p *Provider) Estimate(ctx context.Context, assetScanStats apitypes.AssetSc
 	return ret, nil
 }
 
-// nolint:cyclop
-func (p *Provider) DiscoverAssets(ctx context.Context) provider.AssetDiscoverer {
-	assetDiscoverer := provider.NewSimpleAssetDiscoverer()
-
-	go func() {
-		defer close(assetDiscoverer.OutputChan)
-
-		regions, err := p.ListAllRegions(ctx)
-		if err != nil {
-			assetDiscoverer.Error = fmt.Errorf("failed to get regions: %w", err)
-			return
-		}
-
-		logger := log.GetLoggerFromContextOrDiscard(ctx)
-
-		for _, region := range regions {
-			instances, err := p.GetInstances(ctx, []ec2types.Filter{}, region.Name)
-			if err != nil {
-				logger.Warnf("Failed to get instances. region=%v: %v", region, err)
-				continue
-			}
-
-			for _, instance := range instances {
-				asset, err := getVMInfoFromInstance(instance)
-				if err != nil {
-					assetDiscoverer.Error = FatalError{
-						Err: fmt.Errorf("failed convert EC2 Instance to AssetType: %w", err),
-					}
-					return
-				}
-				select {
-				case assetDiscoverer.OutputChan <- asset:
-				case <-ctx.Done():
-					assetDiscoverer.Error = ctx.Err()
-					return
-				}
-			}
-		}
-	}()
-
-	return assetDiscoverer
-}
-
 // nolint:nilnil
 func (p *Provider) getInstanceWithID(ctx context.Context, id string, region string) (*ec2types.Instance, error) {
 	out, err := p.ec2Client.DescribeInstances(ctx, &ec2.DescribeInstancesInput{
@@ -746,164 +703,4 @@ func (p *Provider) RemoveAssetScan(ctx context.Context, config *provider.ScanJob
 	}
 	err = errors.Join(listOfErrors...)
 	return err
-}
-
-func (p *Provider) GetInstances(ctx context.Context, filters []ec2types.Filter, regionID string) ([]Instance, error) {
-	ret := make([]Instance, 0)
-
-	input := &ec2.DescribeInstancesInput{
-		MaxResults: to.Ptr[int32](maxResults), // TODO what will be a good number?
-	}
-	if len(filters) > 0 {
-		input.Filters = filters
-	}
-
-	out, err := p.ec2Client.DescribeInstances(ctx, input, func(options *ec2.Options) {
-		options.Region = regionID
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to describe instances: %w", err)
-	}
-	ret = append(ret, p.getInstancesFromDescribeInstancesOutput(ctx, out, regionID)...)
-
-	// use pagination
-	// TODO we can make it better by not saving all results in memory. See https://github.com/openclarity/vmclarity/pull/3#discussion_r1021656861
-	for out.NextToken != nil {
-		input := &ec2.DescribeInstancesInput{
-			MaxResults: to.Ptr[int32](maxResults), // TODO what will be a good number?
-			NextToken:  out.NextToken,
-		}
-		if len(filters) > 0 {
-			input.Filters = filters
-		}
-
-		out, err = p.ec2Client.DescribeInstances(ctx, input, func(options *ec2.Options) {
-			options.Region = regionID
-		})
-		if err != nil {
-			return nil, fmt.Errorf("failed to describe instances: %w", err)
-		}
-		ret = append(ret, p.getInstancesFromDescribeInstancesOutput(ctx, out, regionID)...)
-	}
-
-	return ret, nil
-}
-
-func (p *Provider) getInstancesFromDescribeInstancesOutput(ctx context.Context, result *ec2.DescribeInstancesOutput, regionID string) []Instance {
-	logger := log.GetLoggerFromContextOrDiscard(ctx)
-
-	var ret []Instance
-	for _, reservation := range result.Reservations {
-		for _, instance := range reservation.Instances {
-			// Ignore terminated instances they are destroyed and
-			// will be garbage collected by AWS.
-			if instance.State.Name == ec2types.InstanceStateNameTerminated {
-				continue
-			}
-
-			if err := validateInstanceFields(instance); err != nil {
-				logger.Errorf("Instance validation failed. instance id=%v: %v", to.ValueOrZero(instance.InstanceId), err)
-				continue
-			}
-			rootVol, err := getRootVolumeInfo(ctx, p.ec2Client, instance, regionID)
-			if err != nil {
-				logger.Warnf("Couldn't get root volume info. instance id=%v: %v", to.ValueOrZero(instance.InstanceId), err)
-				rootVol = &apitypes.RootVolume{
-					SizeGB:    0,
-					Encrypted: apitypes.RootVolumeEncryptedUnknown,
-				}
-			}
-
-			ret = append(ret, Instance{
-				ID:                  *instance.InstanceId,
-				Region:              regionID,
-				AvailabilityZone:    *instance.Placement.AvailabilityZone,
-				Image:               *instance.ImageId,
-				InstanceType:        string(instance.InstanceType),
-				Platform:            *instance.PlatformDetails,
-				Tags:                getTagsFromECTags(instance.Tags),
-				LaunchTime:          *instance.LaunchTime,
-				VpcID:               *instance.VpcId,
-				SecurityGroups:      getSecurityGroupsIDs(instance.SecurityGroups),
-				RootDeviceName:      to.ValueOrZero(instance.RootDeviceName),
-				RootVolumeSizeGB:    int32(rootVol.SizeGB),
-				RootVolumeEncrypted: rootVol.Encrypted,
-
-				ec2Client: p.ec2Client,
-			})
-		}
-	}
-	return ret
-}
-
-func getRootVolumeInfo(ctx context.Context, client *ec2.Client, i ec2types.Instance, region string) (*apitypes.RootVolume, error) {
-	if i.RootDeviceName == nil || *i.RootDeviceName == "" {
-		return nil, errors.New("RootDeviceName is not set")
-	}
-	logger := log.GetLoggerFromContextOrDiscard(ctx)
-	for _, mapping := range i.BlockDeviceMappings {
-		if to.ValueOrZero(mapping.DeviceName) == to.ValueOrZero(i.RootDeviceName) {
-			if mapping.Ebs == nil {
-				return nil, errors.New("EBS of the root volume is nil")
-			}
-			if mapping.Ebs.VolumeId == nil {
-				return nil, errors.New("volume ID of the root volume is nil")
-			}
-			descParams := &ec2.DescribeVolumesInput{
-				VolumeIds: []string{*mapping.Ebs.VolumeId},
-			}
-
-			describeOut, err := client.DescribeVolumes(ctx, descParams, func(options *ec2.Options) {
-				options.Region = region
-			})
-			if err != nil {
-				return nil, errors.New("failed to describe the root volume")
-			}
-
-			if len(describeOut.Volumes) == 0 {
-				return nil, errors.New("volume list is empty")
-			}
-			if len(describeOut.Volumes) > 1 {
-				logger.WithFields(logrus.Fields{
-					"VolumeId":    *mapping.Ebs.VolumeId,
-					"Volumes num": len(describeOut.Volumes),
-				}).Warnf("Found more than 1 root volume, using the first")
-			}
-
-			return &apitypes.RootVolume{
-				SizeGB:    int(to.ValueOrZero(describeOut.Volumes[0].Size)),
-				Encrypted: encryptedToAPI(describeOut.Volumes[0].Encrypted),
-			}, nil
-		}
-	}
-
-	return nil, errors.New("instance doesn't have a root volume block device mapping")
-}
-
-func encryptedToAPI(encrypted *bool) apitypes.RootVolumeEncrypted {
-	if encrypted == nil {
-		return apitypes.RootVolumeEncryptedUnknown
-	}
-	if *encrypted {
-		return apitypes.RootVolumeEncryptedYes
-	}
-	return apitypes.RootVolumeEncryptedNo
-}
-
-func (p *Provider) ListAllRegions(ctx context.Context) ([]Region, error) {
-	ret := make([]Region, 0)
-	out, err := p.ec2Client.DescribeRegions(ctx, &ec2.DescribeRegionsInput{
-		AllRegions: nil, // display also disabled regions?
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to describe regions: %w", err)
-	}
-
-	for _, region := range out.Regions {
-		ret = append(ret, Region{
-			Name: *region.RegionName,
-		})
-	}
-
-	return ret, nil
 }
